Apps fixs 7.0
