+++ README.md
@@ -1,2 +1,4 @@
 # cloudshell-orch-sandbox
 Package for sandbox orchestration script functions
+
+#### Supported CloudShell version: 7.0
