+++ SandboxOrchestrationPackage/version.txt
@@ -1,1 +1,1 @@
-1.0.0
+7.0.0
