+++ sandbox_scripts/environment/teardown/teardown_script.py
@@ -6,12 +6,10 @@ from cloudshell.helpers.scripts import cloudshell_scripts_helpers as helpers
 from cloudshell.api.common_cloudshell_api import CloudShellAPIError
 from cloudshell.core.logger import qs_logger
 from sandbox_scripts.profiler.env_profiler import profileit
-from sandbox_scripts.helpers.resource_helpers import get_vm_custom_param, get_resources_created_in_res
+from sandbox_scripts.helpers.vm_details_helper import get_vm_custom_param
 
 
 class EnvironmentTeardown:
-    REMOVE_DEPLOYED_RESOURCE_ERROR = 153
-
     def __init__(self):
         self.reservation_id = helpers.get_reservation_context_details().id
         self.logger = qs_logger.get_qs_logger(log_file_prefix="CloudShell Sandbox Teardown",
@@ -28,9 +26,7 @@ class EnvironmentTeardown:
 
         self._disconnect_all_routes_in_reservation(api, reservation_details)
 
-        self._power_off_and_delete_all_vm_resources(api, reservation_details, self.reservation_id)
-
-        self._cleanup_connectivity(api, self.reservation_id)
+        self._power_off_and_delete_all_vm_resources(api, reservation_details)
 
         self.logger.info("Teardown for reservation {0} completed".format(self.reservation_id))
         api.WriteMessageToReservationOutput(reservationId=self.reservation_id,
@@ -46,6 +42,8 @@ class EnvironmentTeardown:
 
         if not endpoints:
             self.logger.info("No routes to disconnect for reservation {0}".format(self.reservation_id))
+            api.WriteMessageToReservationOutput(reservationId=self.reservation_id,
+                                                message="Nothing to disconnecting")
             return
 
         try:
@@ -67,16 +65,8 @@ class EnvironmentTeardown:
             api.WriteMessageToReservationOutput(reservationId=self.reservation_id,
                                                 message="Error disconnecting apps. Error: {0}".format(exc.message))
 
-    def _power_off_and_delete_all_vm_resources(self, api, reservation_details, reservation_id):
-        """
-        :param CloudShellAPISession api:
-        :param GetReservationDescriptionResponseInfo reservation_details:
-        :param str reservation_id:
-        :return:
-        """
-        # filter out resources not created in this reservation
-        resources = get_resources_created_in_res(reservation_details=reservation_details,
-                                                 reservation_id=reservation_id)
+    def _power_off_and_delete_all_vm_resources(self, api, reservation_details):
+        resources = reservation_details.ReservationDescription.Resources
 
         pool = ThreadPool()
         async_results = []
@@ -104,14 +94,7 @@ class EnvironmentTeardown:
 
         # delete resource - bulk
         if resource_to_delete:
-            try:
-                api.RemoveResourcesFromReservation(self.reservation_id, resource_to_delete)
-            except CloudShellAPIError as exc:
-                if exc.code == EnvironmentTeardown.REMOVE_DEPLOYED_RESOURCE_ERROR:
-                    self.logger.error(
-                            "Error executing RemoveResourcesFromReservation command. Error: {0}".format(exc.message))
-                    api.WriteMessageToReservationOutput(reservationId=self.reservation_id,
-                                                        message=exc.message)
+            api.DeleteResources(resource_to_delete)
 
     def _power_off_or_delete_deployed_app(self, api, resource_info, lock, message_status):
         """
@@ -144,9 +127,10 @@ class EnvironmentTeardown:
                                 api.WriteMessageToReservationOutput(reservationId=self.reservation_id,
                                                                     message='Apps are being deleted...')
 
-                # removed call to destroy_vm_only from this place because it will be called from
-                # the server in RemoveResourcesFromReservation
-
+                api.ExecuteResourceConnectedCommand(self.reservation_id,
+                                                    resource_name,
+                                                    "destroy_vm_only",
+                                                    "remote_app_management")
                 return resource_name
             else:
                 power_off = "true"
@@ -174,14 +158,3 @@ class EnvironmentTeardown:
             self.logger.error("Error deleting or powering off deployed app {0} in reservation {1}. Error: {2}"
                               .format(resource_name, self.reservation_id, str(exc)))
             return None
-
-    def _cleanup_connectivity(self, api, reservation_id):
-        """
-        :param CloudShellAPISession api:
-        :param str reservation_id:
-        :return:
-        """
-        self.logger.info("Cleaning-up connectivity for reservation {0}".format(self.reservation_id))
-        api.WriteMessageToReservationOutput(reservationId=self.reservation_id,
-                                            message='Cleaning-up connectivity')
-        api.CleanupSandboxConnectivity(reservation_id)
